\name{coxScoreHess}
\alias{coxScoreHess}

\title{Calculate the Score vector / Hessian matrix for the Cox model}
\description{Calculate the partial log Likelihood, Score vector or the Hessian matrix for the Cox proportional hazards model with inputs of covariates, survival outcomes and the relative risks}
\usage{
    coxScoreHess(X, y, exb, hess = FALSE, sorted = FALSE)
}
\arguments{
    \item{X}{the covariate matrix from model.matrix, without the interecpt term.}
    \item{y}{y is a survival object, y = Surv(time, event).}
    \item{exb}{exb is the relative risks with exb = exp(X*beta).}
    \item{hess}{output the Hessian matrix, with hess = FALSE as the default, which outputs the score vector only.}
    \item{sorted}{data were sorted by time from the largest to the smallest, to speed up the algorithm, default is sorted = FALSE, sort by time is recommand when the function will be called multiple times for the same y.}
}
\value{An p by 1 vector of the score of the function calculated at the 
       point relative exp(X*beta). If hess = TRUE, then a list with the following three components is returned:

    \item{score}{a 1 x p score vector.}
    \item{Sigma}{a p x p matrix for the empirical varaince of the score.}
    \item{H}{a p x p hessian matrix.}
}

\details{The survival time shall be sorted from the largest to the smallest, an error will occur if y is not sorted.

     partial likelihood = sum(event(exp(X*beta)/S0))

     score = sum(event*(X - S1/S0))

     Sigma = sum(S1*t(S1))

     H     = sum(event*(S2/S0 - S1*t(S1)/S0))

  the robust varaince can be calculated by inv(H)*Sigma*inv(H).
}

\seealso{
    \code{\link{numHessian}}
    \code{\link{numScore}}
    \code{\link{multiRoot}}
}

%this is a simple function, no example is needed here.
\examples{
  n = 5
  x = matrix(runif(n*3), nrow = n)
  y = Surv(runif(n), rbinom(n, 1, 0.7))
  ### offset term
  ot= runif(n)
  ### fit the coxph model
  #fit = coxph(y~x+offset(ost))
  #beta = fit$coefficientspartial 
  beta = c(0.3, 0.4, 0.5)
  #cat("The Cox partial likelihood for the fitted coxph model:\n")
  #print(logLik(fit))
  cat("The Cox partial likelihood using the coxlogLik function:\n", 
    coxlogLik(x, y, beta, offset = ot), "\n")

}
